demo: Render the Article JSON-LD from a self-contained component
The Article structured data was split across a `buildArticle` helper, a separate
image-URL helper and the JSX that wired them into the page. `OrganizationJsonLd`
already shows the better shape: one component that carries its own fragment and
renders itself.

Turn the Article into `NewsArticleJsonLd` accordingly and group it by domain
(news) rather than by type, so `util/structuredData` disappears. The image-URL
resolution is private to the component, since the aspect ratios it produces are
specific to the Article image. Emitted output is unchanged.
